feat: 修改webview和JS介面互動寫法 webview新功能撰寫 分享功能/登出/未讀顯示 首頁邏輯修改 API部分修改

+++ app/src/main/java/com/taiwanlife/teamwalk/remote/response/api/UserInfoResponse.kt
@@ -147,7 +147,7 @@ data class UserInfoResponse(
      * BMI值，由後端直接計算(首頁使用)
      */
     @SerializedName("bmi")
-    val bmi: Int?,
+    val bmi: Float?,
     /**
      * 連續簽到天數(我的頁面使用)
      */
